frontend: helpers: Extract debugVerbose to own module
